split plugins test
